Fix issue with statistics (passenger miles, total miles). - The returning type of TotalNumberOfMiles() was uint16_t instead of float. - To calculate the total passenger miles the multiplication was taking the aircraft passengers instead of the total number of passengers. - Ordered statistics prints to make easier reviewing them.

+++ aircrafts/AircraftType.h
@@ -144,7 +144,7 @@ public:
      * 
      * @return The total number of miles. 
      */
-    inline uint16_t TotalNumberOfMiles() const { return mfTotalNumberOfMiles; }
+    inline float TotalNumberOfMiles() const { return mfTotalNumberOfMiles; }
 
     /**
      * @brief Get the total number of passengers transported by the aircrafts of this type.
